django-course/taskflow/requirements.txt
asgiref==3.8.1
coverage==7.6.10
Django==5.1.5
django-cors-headers==4.6.0
django-formtools==2.5.1
django-jazzmin==3.0.1
django-redis==5.4.0
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pillow==11.1.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
whitenoise==6.8.2
